utils: move FdExt tests to right module
